Set the MFA status page name.
Removed success_url from MfaEnableView, use success_page instead.

Check if an MfaModel object already exists for a user when processing
a GET or POST request for MfaEnableView.  Redirect to the MFA status
page if it does.

Set mfa_enabled to True in the user's session when MFA is successfully
enabled and save an MfaModel to the database.

Added the MfaEnableSuccessView class.  Show a success page when MFA is
enabled.  Redirect to the MFA status page if the page is refreshed or
accessed directly.

Replaced success_url with success_page in the MfaDisableView class.

Check that an MfaModel record exists for a user when processing a GET or
POST request for MfaDisableView.  Redirect to the MFA status page if it
does not.

Delete the user's MfaModel object on a successful POST to the
MfaDisableView class and redirect to the success page.

Added the MfaDisableSuccessView class.  Redirect to the MFA status page
if the success page is refreshed or accessed directly.