Set a top-level status_page variable.
PasswordChangeView should also inherit from LoginRequiredMixin.

Specify form class and page_name attributes for the PasswordChangeView
class.

Set a response_dict attribute when creating instances of
PasswordChangeView.

Implement the get method for PasswordChangeView and return the
instance's response_dict when rendering the page template.

Redirect to a success page for a successful post to PasswordChangeView.

Added a PasswordChangeSuccessView class.  Redirect to the status page if
it is refreshed or accessed directly.