Set login_page at the top-level.
ActivateRegistrationView -> RegistrationConfirmView

Added page_name, success_page and fail_page attributes to
RegistrationView.

Implement __init__() and get() for RegistrationView.

Replaced form.register() with form.save() in post() for
RegistrationView.

Redirect to success or fail pages in the post() method of
RegistrationView instead of rendering the page template.

Added the RegistrationSuccessView and RegistrationFailedView classes.

Added the page_name, success_page, fail_page and form_class attributes
to RegistrationConfirmView.

Set template_name to None in RegistrationConfirmView since we don't
render a page; we just redirect to the success or fail pages.

Added __init__() and get() methods to RegistrationConfirmView.

Moved logic from get() in RegistrationConfirmView to the
RegistrationConfirmForm class.

Added the RegistrationConfirmSuccessView and
RegistrationConfirmFailedView classes.

Replaced hardcoded 'login' path with a reference to login_page in
SendPasswordResetView.