Replaced the user_already_registered error with two new errors, user_already_exists and user_already_activated in RegistrationForm.
Look up the activation record and assign it to the form if a user record
is located during RegistrationForm.clean_username().  This is needed to
allow the registration email to be re-sent.

Renamed the username_invalid error to user_invalid in
RegistrationConfirmForm.

Renamed the username_already_active error to user_already_exists in
RegistrationConfirmForm.