Suppress remaining form errors in RegistrationView.post() if a user_already_activated error is present.
Add a query_string containing an activation_id parameter when
redirecting to the fail page in RegistrationView.post() if a user
already exists, but isn't active.

Pass errors to the fail page in RegistrationView.post() using Django
messages.

Suppress remaining form errors in RegistrationConfirmView.get() if a
user_already_exists error is present.

Add a query_string containing an activation_id parameter when
redirecting to the fail page in RegistrationConfirmView.post() if a user
already exists.

Pass errors to the fail page in RegistrationConfirmForm.post() using Django
messages.